+++ CHANGELOG.md
@@ -4,6 +4,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+Features:
+- Add component `k-PhoneDropdown`.
+- Add new helper `getComputedLeft` on `ElementHelper`.
+- Add text utilities classes: `k-u-small` and `k-u-strong`.
+- Add `k-dropdownArrow` helper.
+
+Breaking changes:
+- Rename `SimpleList` to `NavList`.
+- Rename `k-Dropdown` props: `contentHorizontalPosition`
+and `arrowHorizontalPosition`.
+
 ## [v0.18.0] - 2016-11-14
 
 Features: